// apps/desktop/src/shared/api/generated/ModRequirementDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ModRequirementDto = {
  unique_id: string;
  /**
   * The name of the mod in the profile that has this UniqueID.
   */
  name: string | null;
  installed_version: string | null;
  minimum_version: string | null;
  /**
   * "required", "optional" or "content_pack_for".
   */
  kind: string;
  /**
   * "satisfied", "missing", "disabled" or "too_old".
   */
  status: string;
};